feat: select default billing entity when creating a customer
